requirements.txt
fastapi
uvicorn
pandas
mistralai
python-dotenv